+++ themes/bootstrap/public/mailbox.js
@@ -4,6 +4,13 @@ var mediumDeviceWidth = 980;
 var messageListMargin = 275;
 var clipboard = null;
 var messageListScroll = false;
+var messageListData = null;
+
+// clearMessageSearch resets the message list search
+function clearMessageSearch() {
+  $('#message-search').val('');
+  updateMessageSearch();
+}
 
 // deleteMessage sends a delete request for a message
 function deleteMessage(id) {
@@ -11,7 +18,7 @@ function deleteMessage(id) {
   $.ajax({
     type: 'DELETE',
     url: '/mailbox/' + mailbox + '/' + id,
-    success: reloadList
+    success: loadList
   })
 }
 
@@ -29,9 +36,37 @@ function htmlView(id) {
       'menubar=yes,resizable=yes,scrollbars=yes,status=yes,toolbar=yes');
 }
 
-// loadList displays the message list for this mailbox
+// loadList loads the message list for this mailbox via AJAX
 function loadList() {
-  $('#message-list').load('/mailbox/' + mailbox, onListLoaded);
+  $('#message-list').hide().empty();
+  $.ajax({
+    dataType: "json",
+    url: '/api/v1/mailbox/' + mailbox,
+    success: function(data) {
+      messageListData = data;
+      // Render list
+      $('#message-list').loadTemplate($('#list-entry-template'), data);
+      $('.message-list-entry').click(onMessageListClick);
+      // Reveal and select current message
+      $("#message-list").slideDown();
+      if (selected != "") {
+        $("#" + selected).click();
+        selected = "";
+      }
+      onDocumentChange();
+      updateMessageSearch();
+    }
+  });
+}
+
+// makeDelay creates a call-back timer that prevents itself from being
+// stacked
+function makeDelay(ms) {
+  var timer = 0;
+  return (function(callback) {
+    clearTimeout (timer);
+    timer = setTimeout(callback, ms);
+  });
 }
 
 // messageSource pops open another window for message source
@@ -39,12 +74,6 @@ function messageSource(id) {
   window.open('/mailbox/' + mailbox + '/' + id + "/source", '_blank',
       'width=800,height=600,' +
       'menubar=no,resizable=yes,scrollbars=yes,status=no,toolbar=no');
-}
-
-// reloadList reloads the message list for this mailbox
-function reloadList() {
-  $('#message-list').hide();
-  loadList();
 }
 
 // toggleMessageLink shows/hids the message link URL form
@@ -75,28 +104,37 @@ function onDocumentChange() {
 
 // onDocumentReady is called by mailbox/index.html to initialize
 function onDocumentReady() {
+  // Prevent search and resize handlers being called too often
+  var searchDelay = makeDelay(200);
+  var resizeDelay = makeDelay(100);
+  $.addTemplateFormatter({
+    "date": function(value, template) {
+      return moment(value).calendar();
+    },
+    "subject": function(value, template) {
+      if (value == null || value.length == 0) {
+        return "(No Subject)";
+      }
+      return value;
+    }
+  });
   $("#message-list").hide();
   onWindowResize();
-  $(window).resize(onWindowResize);
+  $(window).resize(function() {
+    resizeDelay(onWindowResize);
+  });
+  $('#message-search').on('change keyup', function(el) {
+    searchDelay(updateMessageSearch);
+  });
   loadList();
 }
 
-// onListLoaded is called when the message list changes
-function onListLoaded() {
-  onDocumentChange();
-  $('.listEntry').click(
-      function() {
-        $('.listEntry').removeClass("disabled");
-        $(this).addClass("disabled");
-        $('#message-content').load('/mailbox/' + mailbox + '/' + this.id, onMessageLoaded);
-        selected = this.id;
-      }
-      )
-    $("#message-list").slideDown();
-  if (selected != "") {
-    $("#" + selected).click();
-    selected = "";
-  }
+// onMessageListClick is triggered by clicks on the message list
+function onMessageListClick() {
+  $('.message-list-entry').removeClass("disabled");
+  $(this).addClass("disabled");
+  $('#message-content').load('/mailbox/' + mailbox + '/' + this.id, onMessageLoaded);
+  selected = this.id;
 }
 
 // onMessageLoaded is called each time a new message is shown
@@ -113,15 +151,38 @@ function onMessageLoaded(responseText, textStatus, XMLHttpRequest) {
 // onWindowResize handles special cases when window is resized
 function onWindowResize() {
   if ($(window).width() > mediumDeviceWidth) {
+    var content_height = $(window).height() - messageListMargin;
+    var messageListWrapper = $('#message-list-wrapper');
+    messageListWrapper.height(content_height);
     if (!messageListScroll) {
-      messageListScroll= true;
-      var content_height = $(window).height() - messageListMargin;
-      $('#message-list-wrapper').height(content_height).addClass("message-list-scroll");
+      messageListScroll = true;
+      messageListWrapper.addClass("message-list-scroll");
     }
   } else {
     if (messageListScroll) {
-      messageListScroll= false;
+      messageListScroll = false;
       $('#message-list-wrapper').height('auto').removeClass("message-list-scroll");
+    }
+  }
+}
+
+// updateMessageSearch compares the message list subjects and senders against
+// the search string and hides entries that don't match
+function updateMessageSearch() {
+  var criteria = $('#message-search').val();
+  if (criteria.length < 2) {
+    $('.message-list-entry').show();
+    return;
+  }
+  criteria = criteria.toLowerCase();
+  for (i=0; i<messageListData.length; i++) {
+    entry = messageListData[i];
+    if ((entry.subject.toLowerCase().indexOf(criteria) > -1) ||
+        (entry.from.toLowerCase().indexOf(criteria) > -1)) {
+      // Match
+      $('#' + entry.id).show();
+    } else {
+      $('#' + entry.id).hide();
     }
   }
 }